requirements.txt
pdfplumber==0.10.2
pdfminer.six==20221105
